Create Set Matrix Zeroes
Write a program to modify a given m x n matrix. If an element in the matrix is 0, set its entire row and column to 0. The input matrix will contain integers and is guaranteed to be at least 1x1 in size.

Input Format:
The input starts with two integers, m and n, representing the number of rows and columns in the matrix.
The next m * n integers correspond to the elements of the matrix in row-wise order.
Output Format:
Print the modified matrix, row by row, with elements separated by a single space.
[Refer sample input and output]
[All text in bold refers to the input]
Sample Input 1:
Enter the number of row:

3

Enter the number of coloumn:

4

The Matrix is:

1 2 3 4 

5 0 6 7 

8 9 10 11 

Sample Output 1:
The Resultant Matrix is:

1 0 3 4 

0 0 0 0 

8 0 10 11