Create Add two Lists using Linked Lists
Write a program to add two non-negative integers represented as linked lists. The digits are stored in reverse order, and each node contains a single digit. The function should return the sum as a linked list in the same reverse order format.

Input Format:
The first input line contains an integer n, the number of nodes in the first linked list.
The second input line contains n integers, representing the digits of the first number.
The third input line contains an integer m, the number of nodes in the second linked list.
The fourth input line contains m integers, representing the digits of the second number.
Output Format:
Print the resulting linked list representing the sum, with each digit separated by a space.
[Refer sample input and output]
[All text in bold refers to the input]

Sample Input :

Enter the element for List 1:

3

2 4 3

Enter the element for List 2:

3

5 6 4

Sample Output :

The List is:

7 0 8