Create Remove Duplicates from Sorted Array
Write a program to remove duplicates from a sorted array of integers and find the length of the new array without duplicates. The program should also print the modified array containing only the unique elements.

Input and Output Format
Input consists of N+1 integers separated by spaces, where the first integer N represents the number of elements in the array, and the following N integers represent the elements of the array. The array is sorted in non-decreasing order. Assume that the maximum value of N is 20.

Output consists of two lines:

The first line contains a single integer representing the length of the array after removing duplicates.
The second line contains the modified array with only unique elements, separated by spaces.

Refer Sample Input and Output.

[All text in bold corresponds to input and the rest corresponds to output]


Sample Input and Output
Sample Input :

Enter the number of elements in the array:
6
Enter the array elements:
1 1 2 2 3 3


Sample Output :

The length of the array after removing duplicates is 3
The modified array is 1 2 3