Create Container with most water
In the ancient city of Aquatica, there stands the Fountain of Vitality, said to hold the power to grant eternal life. Legend has it that the fountain's basin has the shape of a rectangular container, filled with mystical water. The amount of vitality the fountain can bestow upon a seeker is determined by the amount of water it can contain.

Your task is to write a C program to find the container with the most water, maximizing the vitality it can hold within the Fountain of Vitality.

Input and Output Format:

Input consists of N+1 lines. The first line contains a single integer N representing the number of elements in the array. The following N lines contain the heights of the walls of the containers, each height separated by a space.

Output consists of a single integer representing the maximum amount of water the container can hold.

Refer Sample Input and Output.

[All text in bold corresponds to input and the rest corresponds to output]

Sample Input :

Enter the number of walls of the container: 4

Enter the heights of the walls of the container: 1 8 6 2

Sample Output :

The maximum amount of water the container can hold is: 6