Create Middle of the Linked List
In the land of DataStructura, the king has a linked list of integers. He is curious to know the middle element of the list. The middle element is defined as the element at position n/2 (where n is the number of nodes in the list). If the number of nodes is even, return the second middle element.

Your task is to help the king by writing a program to find the middle of the linked list.

Structure Definition:
typedef struct ListNode {
    int val;
    struct ListNode* next;
} ListNode;
 

Include Functions:
ListNode* createLinkedList(int* values, int size); - Creates a linked list from the provided array of values.
ListNode* findMiddle(ListNode* head); - Finds and returns the middle node of the linked list.
void displayMiddle(ListNode* middle); - Displays the value of the middle node.
Constraints:
Number of nodes: [1, 100]
Node values: [-1000, 1000]
Input and Output Format:
Input:

An integer n representing the number of elements in the linked list.
A list of n  node values as space-separated integers.
Output:

The value of the middle node of the linked list.
Refer to the sample input and output for formatting specifications.

[All text in bold corresponds to input and the rest corresponds to output]

Sample Input and Output 1:
Enter the number of nodes:

5

Enter node values:

1 2 3 4 5

The middle node's value is:

3

Sample Input and Output 2:
Enter the number of nodes:

6

Enter node values:

10 20 30 40 50 60

The middle node's value is:

40