Create Valid Palindrome
Given a string s, determine if it is a palindrome. A phrase is considered a palindrome if, after converting all uppercase letters into lowercase letters and removing all non-alphanumeric characters, it reads the same forward and backward. Alphanumeric characters include letters and numbers.

Input Format:
The input consists of a single line containing the string s.

Output Format:
Print true if the string is a palindrome, otherwise print false.

[Refer sample input and output]
[All text in bold refers to the input]

Sample input & output 1:
Enter the string: race a car
false

Sample input & output 2:
Enter the string: Was it a car or a cat I saw?
true

Sample input & output 3:
Enter the string: 1
true