Create Valid Number
Write a program to determine whether a given string s is a valid number. A valid number can be defined as either an integer or a decimal number, optionally followed by an exponent. An integer number can have an optional sign and must contain digits. A decimal number can have an optional sign and must follow one of several formats involving digits and a decimal point. An exponent is represented by an 'e' or 'E' followed by an integer.

Input and Output Format:
Input consists of a string s. Assume that the maximum length of the string is 1000.
Output is true if the string s represents a valid number, and false otherwise.
Refer to the sample input and output for formatting specifications.

All text in bold corresponds to input and the rest corresponds to output.

Sample Input and Output 1:
Enter the string : 0
The input string is "0"
The string is a valid number: true
Sample Input and Output 2:
Enter the string : e
The input string is "e"
The string is a valid number: false